refactor(sidecars): extract shared gRPC infrastructure
