+++ lib/backend-common/Cargo.toml
@@ -40,7 +40,6 @@ tokio = { workspace = true }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
 tracing-opentelemetry = { workspace = true }
-url = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
