+++ lib/backend-common/src/args.rs
@@ -7,62 +7,11 @@
 //! the same CLI surface across Rust and Python backends. Engines extend this
 //! with their own `clap` `Args` using `#[command(flatten)]`.
 
-use std::num::NonZeroUsize;
 use std::path::PathBuf;
-use std::time::Duration;
 
 use clap::Args;
 
 use crate::disagg::DisaggregationMode;
-
-const DEFAULT_GRPC_CONNECT_ATTEMPT_TIMEOUT_SECS: u64 = 30;
-const DEFAULT_GRPC_CONNECTIONS: NonZeroUsize = NonZeroUsize::new(8).unwrap();
-const DEFAULT_GRPC_RETRY_INTERVAL_SECS: u64 = 1;
-const DEFAULT_GRPC_STARTUP_DEADLINE_SECS: u64 = 300;
-
-/// Normalize a plaintext gRPC client endpoint for tonic.
-///
-/// Bare addresses and `grpc://` URLs are translated to `http://`. TLS and
-/// other schemes are rejected until the caller configures tonic TLS.
-pub fn normalize_grpc_endpoint(raw: &str, argument: &str) -> Result<String, String> {
-    let endpoint = raw.trim();
-    if endpoint.is_empty() {
-        return Err(format!(
-            "`{argument}` is required and must specify a gRPC server address, such as `http://HOST:PORT`"
-        ));
-    }
-
-    let normalized = if let Some(authority) = endpoint.strip_prefix("grpc://") {
-        format!("http://{authority}")
-    } else if endpoint.starts_with("http://") {
-        endpoint.to_string()
-    } else if endpoint.starts_with("grpcs://") || endpoint.starts_with("https://") {
-        return Err(format!("TLS endpoints are not supported by `{argument}`"));
-    } else if endpoint.contains("://") {
-        return Err(format!(
-            "unsupported endpoint scheme in `{argument}`: `{endpoint}`"
-        ));
-    } else {
-        format!("http://{endpoint}")
-    };
-
-    let parsed = url::Url::parse(&normalized)
-        .map_err(|error| format!("invalid gRPC endpoint for `{argument}`: {error}"))?;
-    if parsed.host().is_none() {
-        return Err(format!("`{argument}` must include a host"));
-    }
-    if !parsed.username().is_empty() || parsed.password().is_some() {
-        return Err(format!("`{argument}` must not include user information"));
-    }
-    if parsed.path() != "/" || parsed.query().is_some() || parsed.fragment().is_some() {
-        return Err(format!(
-            "`{argument}` must contain only a plaintext scheme and authority"
-        ));
-    }
-
-    let authority = &parsed[url::Position::BeforeHost..url::Position::AfterPort];
-    Ok(format!("http://{authority}"))
-}
 
 /// CLI flags that every Rust backend needs. Flatten these into your engine's
 /// `Args` struct:
@@ -141,112 +90,4 @@ pub struct CommonArgs {
     /// shim does not read this env var.
     #[arg(long, default_value_t = false, env = "DYN_ROUTE_TO_ENCODER")]
     pub route_to_encoder: bool,
-
-    /// Number of parallel sidecar gRPC connections. The default of eight was
-    /// sufficient in sidecar load tests to avoid connection-level throttling
-    /// at high request concurrency.
-    #[arg(
-        long = "grpc-connections",
-        env = "DYN_SIDECAR_GRPC_CONNECTIONS",
-        default_value_t = DEFAULT_GRPC_CONNECTIONS
-    )]
-    pub grpc_connections: NonZeroUsize,
-
-    /// Maximum duration of one sidecar gRPC connection attempt.
-    #[arg(
-        long = "grpc-connect-attempt-timeout-secs",
-        env = "DYN_SIDECAR_GRPC_CONNECT_ATTEMPT_TIMEOUT_SECS",
-        default_value_t = DEFAULT_GRPC_CONNECT_ATTEMPT_TIMEOUT_SECS,
-        value_parser = clap::value_parser!(u64).range(1..)
-    )]
-    pub grpc_connect_attempt_timeout_secs: u64,
-
-    /// Delay between sidecar gRPC connection attempts.
-    #[arg(
-        long = "grpc-retry-interval-secs",
-        env = "DYN_SIDECAR_GRPC_RETRY_INTERVAL_SECS",
-        default_value_t = DEFAULT_GRPC_RETRY_INTERVAL_SECS,
-        value_parser = clap::value_parser!(u64).range(1..)
-    )]
-    pub grpc_retry_interval_secs: u64,
-
-    /// Maximum total duration for sidecar gRPC startup.
-    #[arg(
-        long = "grpc-startup-deadline-secs",
-        env = "DYN_SIDECAR_GRPC_STARTUP_DEADLINE_SECS",
-        default_value_t = DEFAULT_GRPC_STARTUP_DEADLINE_SECS,
-        value_parser = clap::value_parser!(u64).range(1..)
-    )]
-    pub grpc_startup_deadline_secs: u64,
-}
-
-impl CommonArgs {
-    pub fn grpc_transport_config(&self) -> GrpcTransportConfig {
-        GrpcTransportConfig {
-            connections: self.grpc_connections,
-            connect_attempt_timeout: Duration::from_secs(self.grpc_connect_attempt_timeout_secs),
-            retry_interval: Duration::from_secs(self.grpc_retry_interval_secs),
-            startup_deadline: Duration::from_secs(self.grpc_startup_deadline_secs),
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct GrpcTransportConfig {
-    pub connections: NonZeroUsize,
-    pub connect_attempt_timeout: Duration,
-    pub retry_interval: Duration,
-    pub startup_deadline: Duration,
-}
-
-impl Default for GrpcTransportConfig {
-    fn default() -> Self {
-        Self {
-            connections: DEFAULT_GRPC_CONNECTIONS,
-            connect_attempt_timeout: Duration::from_secs(DEFAULT_GRPC_CONNECT_ATTEMPT_TIMEOUT_SECS),
-            retry_interval: Duration::from_secs(DEFAULT_GRPC_RETRY_INTERVAL_SECS),
-            startup_deadline: Duration::from_secs(DEFAULT_GRPC_STARTUP_DEADLINE_SECS),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::normalize_grpc_endpoint;
-
-    const ARGUMENT: &str = "--test-endpoint";
-
-    #[test]
-    fn normalizes_plaintext_grpc_endpoints() {
-        assert_eq!(
-            normalize_grpc_endpoint(" 127.0.0.1:50051 ", ARGUMENT).unwrap(),
-            "http://127.0.0.1:50051"
-        );
-        assert_eq!(
-            normalize_grpc_endpoint("http://server:50051", ARGUMENT).unwrap(),
-            "http://server:50051"
-        );
-        assert_eq!(
-            normalize_grpc_endpoint("grpc://server:50051", ARGUMENT).unwrap(),
-            "http://server:50051"
-        );
-    }
-
-    #[test]
-    fn rejects_missing_hosts_and_unsupported_schemes() {
-        for endpoint in [
-            "",
-            " ",
-            "http://",
-            "grpc://",
-            "https://server",
-            "other://server",
-            "http://user:password@server:50051",
-            "http://server:50051/path",
-            "http://server:50051?token=secret",
-            "http://server:50051#fragment",
-        ] {
-            assert!(normalize_grpc_endpoint(endpoint, ARGUMENT).is_err());
-        }
-    }
 }
