+++ lib/backend/sidecar-common/Cargo.toml
@@ -1,0 +1,22 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "dynamo-sidecar-common"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Shared gRPC transport and CLI infrastructure for Dynamo sidecars"
+
+[dependencies]
+dynamo-backend-common = { workspace = true }
+
+clap = { version = "4", features = ["derive", "env"] }
+futures = { workspace = true }
+tokio = { workspace = true }
+tonic = { workspace = true }
+tracing = { workspace = true }
+url = { workspace = true }
