+++ lib/backend/sidecar-common/src/error.rs
@@ -1,0 +1,72 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use dynamo_backend_common::{BackendError, DynamoError, ErrorType};
+
+fn backend(kind: BackendError, message: impl Into<String>) -> DynamoError {
+    DynamoError::builder()
+        .error_type(ErrorType::Backend(kind))
+        .message(message)
+        .build()
+}
+
+pub fn invalid_argument(message: impl Into<String>) -> DynamoError {
+    backend(BackendError::InvalidArgument, message)
+}
+
+pub fn protocol_error(peer: &str, message: impl Into<String>) -> DynamoError {
+    backend(
+        BackendError::Unknown,
+        format!("invalid {peer} gRPC response: {}", message.into()),
+    )
+}
+
+pub fn engine_shutdown(message: impl Into<String>) -> DynamoError {
+    backend(BackendError::EngineShutdown, message)
+}
+
+pub fn cannot_connect(message: impl Into<String>) -> DynamoError {
+    backend(BackendError::CannotConnect, message)
+}
+
+pub fn status_to_dynamo(rpc: &str, status: tonic::Status) -> DynamoError {
+    let kind = match status.code() {
+        tonic::Code::InvalidArgument
+        | tonic::Code::NotFound
+        | tonic::Code::OutOfRange
+        | tonic::Code::FailedPrecondition
+        | tonic::Code::AlreadyExists => BackendError::InvalidArgument,
+        tonic::Code::Unavailable => BackendError::CannotConnect,
+        tonic::Code::Cancelled => BackendError::Cancelled,
+        tonic::Code::DeadlineExceeded => BackendError::ConnectionTimeout,
+        _ => BackendError::Unknown,
+    };
+    backend(
+        kind,
+        format!("{rpc}: {} ({:?})", status.message(), status.code()),
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use dynamo_backend_common::{BackendError, ErrorType};
+
+    use super::status_to_dynamo;
+
+    #[test]
+    fn maps_transport_statuses_to_backend_errors() {
+        for (code, expected) in [
+            (tonic::Code::InvalidArgument, BackendError::InvalidArgument),
+            (tonic::Code::Unavailable, BackendError::CannotConnect),
+            (tonic::Code::Cancelled, BackendError::Cancelled),
+            (
+                tonic::Code::DeadlineExceeded,
+                BackendError::ConnectionTimeout,
+            ),
+            (tonic::Code::Internal, BackendError::Unknown),
+        ] {
+            let error = status_to_dynamo("Test", tonic::Status::new(code, "failure"));
+            assert_eq!(error.error_type(), ErrorType::Backend(expected));
+        }
+    }
+}
