+++ lib/backend/vllm-sidecar/Cargo.toml
@@ -21,6 +21,7 @@ path = "src/main.rs"
 
 [dependencies]
 dynamo-backend-common = { workspace = true }
+dynamo-sidecar-common = { workspace = true }
 
 anyhow = { workspace = true }
 async-stream = { workspace = true }
