+++ lib/backend/vllm-sidecar/src/tests.rs
@@ -9,9 +9,10 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 
 use dynamo_backend_common::{
-    DisaggregationMode, FinishReason, GenerateContext, GrpcTransportConfig, LLMEngine,
-    OutputOptions, PrefillResult, PreprocessedRequest, SamplingOptions, StopConditions,
+    DisaggregationMode, FinishReason, GenerateContext, LLMEngine, OutputOptions, PrefillResult,
+    PreprocessedRequest, SamplingOptions, StopConditions,
 };
+use dynamo_sidecar_common::{GrpcEndpoint, GrpcTransportConfig};
 use futures::{Stream, StreamExt};
 use serde_json::json;
 use tokio::net::TcpListener;
@@ -397,7 +398,7 @@ fn request() -> PreprocessedRequest {
 
 fn engine(endpoint: &str, mode: DisaggregationMode, connections: usize) -> VllmSidecarEngine {
     VllmSidecarEngine::new(
-        endpoint.to_string(),
+        GrpcEndpoint::parse(endpoint, "--vllm-endpoint").expect("valid test endpoint"),
         ConfiguredModel {
             source: "model-source".to_string(),
         },
@@ -535,7 +536,8 @@ async fn pool_uses_each_configured_connection() {
         connections: NonZeroUsize::new(2).unwrap(),
         ..Default::default()
     };
-    let client = VllmClient::connect(&server.endpoint, transport)
+    let endpoint = GrpcEndpoint::parse(&server.endpoint, "--vllm-endpoint").unwrap();
+    let client = VllmClient::connect(&endpoint, transport)
         .await
         .expect("connect pool");
     assert_eq!(client.connection_count(), 2);
@@ -561,36 +563,6 @@ async fn pool_uses_each_configured_connection() {
         .map(SocketAddr::port)
         .collect();
     assert_eq!(ports.len(), 2);
-}
-
-#[tokio::test]
-async fn startup_deadline_caps_connection_retries() {
-    let listener = TcpListener::bind("127.0.0.1:0").await.expect("bind");
-    let address = listener.local_addr().expect("address");
-    drop(listener);
-
-    let transport = GrpcTransportConfig {
-        connections: NonZeroUsize::new(2).unwrap(),
-        connect_attempt_timeout: std::time::Duration::from_millis(50),
-        retry_interval: std::time::Duration::from_millis(10),
-        startup_deadline: std::time::Duration::from_millis(100),
-    };
-    let endpoint = format!("http://{address}");
-    let result = tokio::time::timeout(
-        std::time::Duration::from_millis(300),
-        VllmClient::connect(&endpoint, transport),
-    )
-    .await
-    .expect("connection retries must respect the startup deadline");
-
-    let error = match result {
-        Ok(_) => panic!("the endpoint is closed"),
-        Err(error) => error,
-    };
-    assert!(
-        error.to_string().contains("within 100ms"),
-        "unexpected error: {error}"
-    );
 }
 
 #[tokio::test]
